the -check-variables option has been removed from 0.12
Now the validator behaves as if this were `false`.
See: https://github.com/hashicorp/terraform/issues/21408#issuecomment-495259124

However, as a result of my investigation, the validator uses the source
module to check whether the syntax is valid and types/values are being
used correctly.

To react with the changes, this commit adds PWD=directory as an
temporary environment variable on the command execution, because validate
command seems to refer to the current directory.